#491: Generalize KryoWorldSerializationManagerTest.
